+++ src/components/formularios/index.ts
@@ -1,1 +1,2 @@
 export { default as AfectadosForm } from "./AfectadosForm.vue";
+export { default as VinculadosForm } from "./VinculadosForm.vue";
